1、RedisUtil改为RedisOperator，并放入组件包内。redis的key使用冒号层次化。 2、加入RabbitMQ，异步更新浏览量。经测试，还不如不加。 3、PageResult类型参数化，方便反序列化。 4、博客详情使用string缓存，pv使用hash，评论和系统配置使用string缓存。

+++ src/main/java/com/wyy/myblog/dao/BlogCommentMapper.java
@@ -16,9 +16,9 @@ public interface BlogCommentMapper {
 
     List<BlogComment> getBlogCommentsList(PageQuery pageQuery);
 
-    int batchAudit(Integer[] ids);
+    int batchAudit(Long[] ids);
 
-    int batchDelete(Integer[] ids);
+    int batchDelete(Long[] ids);
 
     BlogComment selectByPrimaryKey(Long commentId);
